admin section in urls

+++ routes/web.php
@@ -6,30 +6,41 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', 'App\Http\Controllers\HomeController@index')->name('home.index');
 
-Route::get('/errors/nonexistent', 'App\Http\Controllers\ErrorController@nonexistent')->name('error.nonexistent');
-
-Route::get('/reviews', 'App\Http\Controllers\ReviewController@index')->name('review.index');
-Route::get('/reviews/create', 'App\Http\Controllers\ReviewController@create')->name('review.create');
-Route::post('/reviews/save', 'App\Http\Controllers\ReviewController@save')->name('review.save');
-Route::get('/reviews/{id}', 'App\Http\Controllers\ReviewController@show')->name('review.show');
-Route::delete('/reviews/{id}', 'App\Http\Controllers\ReviewController@destroy')->name('review.destroy');
-
-Route::get('/categories', 'App\Http\Controllers\CategoryController@index')->name('category.index');
-Route::get('/categories/create', 'App\Http\Controllers\CategoryController@create')->name('category.create');
-Route::post('/categories/save', 'App\Http\Controllers\CategoryController@save')->name('category.save');
-Route::get('/categories/{id}', 'App\Http\Controllers\CategoryController@show')->name('category.show');
-Route::delete('/categories/{id}', 'App\Http\Controllers\CategoryController@destroy')->name('category.destroy');
-
-Route::get('/custom-users', 'App\Http\Controllers\CustomUserController@index')->name('custom-user.index');
-Route::get('/custom-users/create', 'App\Http\Controllers\CustomUserController@create')->name('custom-user.create');
-Route::post('/custom-users/save', 'App\Http\Controllers\CustomUserController@save')->name('custom-user.save');
-Route::get('/custom-users/{id}', 'App\Http\Controllers\CustomUserController@show')->name('custom-user.show');
-Route::delete('/custom-users/{id}', 'App\Http\Controllers\CustomUserController@destroy')->name('custom-user.destroy');
-
+# ========================== USER =================================
+# Games
 Route::get('/games', 'App\Http\Controllers\GameController@index')->name('game.index');
-Route::get('/games/create', 'App\Http\Controllers\GameController@create')->name('game.create');
-Route::post('/games/save', 'App\Http\Controllers\GameController@save')->name('game.save');
 Route::get('/games/shopping-cart', 'App\Http\Controllers\GameController@shoppingCart')->name('game.shoppingCart');
 Route::post('/games/add-to-cart/{id}', 'App\Http\Controllers\GameController@addToShoppingCart')->name('game.addToShoppingCart');
 Route::get('/games/{id}', 'App\Http\Controllers\GameController@show')->name('game.show');
-Route::delete('/games/{id}', 'App\Http\Controllers\GameController@destroy')->name('game.destroy');
+
+# ========================== ADMIN ================================
+# Games
+Route::get('/admin/games', 'App\Http\Controllers\GameController@index')->name('admin-game.index');
+Route::get('/admin/games/create', 'App\Http\Controllers\GameController@create')->name('admin-game.create');
+Route::post('/admin/games/save', 'App\Http\Controllers\GameController@save')->name('admin-game.save');
+Route::get('/admin/games/{id}', 'App\Http\Controllers\GameController@show')->name('game.show');
+Route::delete('/admin/games/{id}', 'App\Http\Controllers\GameController@destroy')->name('admin-game.destroy');
+
+# Reviews
+Route::get('/admin/reviews', 'App\Http\Controllers\AdminReviewController@index')->name('admin-review.index');
+Route::get('/admin/reviews/create', 'App\Http\Controllers\AdminReviewController@create')->name('admin-review.create');
+Route::post('/admin/reviews/save', 'App\Http\Controllers\AdminReviewController@save')->name('admin-review.save');
+Route::get('/admin/reviews/{id}', 'App\Http\Controllers\AdminReviewController@show')->name('admin-review.show');
+Route::delete('/admin/reviews/{id}', 'App\Http\Controllers\AdminReviewController@destroy')->name('admin-review.destroy');
+
+# Categories
+Route::get('/admin/categories', 'App\Http\Controllers\AdminCategoryController@index')->name('admin-category.index');
+Route::get('/admin/categories/create', 'App\Http\Controllers\AdminCategoryController@create')->name('admin-category.create');
+Route::post('/admin/categories/save', 'App\Http\Controllers\AdminCategoryController@save')->name('admin-category.save');
+Route::get('/admin/categories/{id}', 'App\Http\Controllers\AdminCategoryController@show')->name('admin-category.show');
+Route::delete('/admin/categories/{id}', 'App\Http\Controllers\AdminCategoryController@destroy')->name('admin-category.destroy');
+
+# Custom Users
+Route::get('/admin/custom-users', 'App\Http\Controllers\AdminCustomUserController@index')->name('admin-custom-user.index');
+Route::get('/admin/custom-users/create', 'App\Http\Controllers\AdminCustomUserController@create')->name('admin-custom-user.create');
+Route::post('/admin/custom-users/save', 'App\Http\Controllers\AdminCustomUserController@save')->name('admin-custom-user.save');
+Route::get('/admin/custom-users/{id}', 'App\Http\Controllers\AdminCustomUserController@show')->name('admin-custom-user.show');
+Route::delete('admin/custom-users/{id}', 'App\Http\Controllers\AdminCustomUserController@destroy')->name('admin-custom-user.destroy');
+
+# ========================== ERRORS =================================
+Route::get('/errors/nonexistent', 'App\Http\Controllers\ErrorController@nonexistent')->name('error.nonexistent');
